requirements.txt
annotated-types==0.7.0
anyio==4.9.0
Authlib==1.6.0
bidict==0.23.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.4
dotenv==0.9.9
exceptiongroup==1.3.0
fastmcp==2.8.1
ffmpeg==1.4
Flask==3.1.1
flask-cors==6.0.1
Flask-SocketIO==5.5.1
google-auth==2.40.3
google-genai==1.21.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mcp==1.9.4
mdurl==0.1.2
numpy==2.3.0
openapi-pydantic==0.5.1
platformdirs==4.3.8
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
python-dotenv==1.1.0
python-engineio==4.12.2
python-multipart==0.0.20
python-socketio==5.13.0
pywhispercpp @ git+https://github.com/absadiki/pywhispercpp@8ecb35b5cf424239fd9fda2d9ca013de1d133397
requests==2.32.4
rich==14.0.0
rsa==4.9.1
shellingham==1.5.4
simple-websocket==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sse-starlette==2.3.6
starlette==0.47.1
tenacity==8.5.0
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.34.3
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
